back-end/pyproject.toml
[project]
name = "back-end"
version = "0.1.0"
description = "Educational Multi-Agent Code Generation System - Backend API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.2.6",
    "anthropic>=0.72.0",
    "fastapi>=0.120.3",
    "openai>=2.6.1",
    "pgvector>=0.4.1",
    "psycopg>=3.2.12",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.0.280",
    "mypy>=1.4.0"
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
]
